Prevent spaces sending .ini
git-svn-id: svn://svn.code.sf.net/p/nikkhokkho/code/trunk@968 ddd8a7e7-04a1-4c0d-ae79-48073a28b335

+++ FileOptimizer/Source/clsUtil.cpp
@@ -930,7 +930,7 @@ int __fastcall clsUtil::Random(int piMin, int piMax)
 
 	
 	iSeed = (214013 * iSeed + 2531011);
-	return((iSeed  % (piMax - piMin)) + piMin);
+	return((iSeed  % (unsigned int) ((piMax - piMin)) + piMin));
 }
